Move CfgWeapons to jam_csla
