+++ addons/jam/jam_csla/config.cpp
@@ -17,3 +17,5 @@ class CfgPatches {
         VERSION_CONFIG;
     };
 };
+
+#include "CfgWeapons.hpp"
